fix: `user_id` column should not be auto-incremented in `personal_access_tokens`

+++ src/db/migrations/1727882863370_create_personal_access_token_table.ts
@@ -4,7 +4,7 @@ export async function up(db: Kysely<any>): Promise<void> {
 	await db.schema
 		.createTable("personal_access_tokens")
 		.addColumn("id", "bigserial", (col) => col.primaryKey())
-		.addColumn("user_id", "bigserial", (col) => col.notNull())
+		.addColumn("user_id", "integer", (col) => col.notNull())
 		.addColumn("fingerprint", "varchar(255)")
 		.addColumn("token", "varchar(64)", (col) => col.unique().notNull())
 		.addColumn("last_used_at", "timestamp")
